requirements.txt
ale_py==0.7.4
apex==0.9.10dev
bitsandbytes==0.35.4
fairscale==0.4.13
gym==0.26.2
huggingface_hub==0.11.1
intel_extension_for_pytorch==1.13.0
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.6.0.66
optuna==3.0.4
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
ray==2.2.0
scikit_learn==1.2.0
scipy==1.9.3
setuptools==65.5.0
stable_baselines3==1.6.2
torch==1.13.0
torch_xla==1.0
tqdm==4.64.1
transformers==4.25.1
wandb==0.13.7
